Simplify the URL API for admin interfaces.
