--- server/auvsi_suas/views/auvsi_admin/__init__.py
@@ -1,1 +1,0 @@
-
