+++ server/auvsi_suas/views/index_test.py
