+++ server/auvsi_suas/views/teams.py
@@ -1,9 +1,11 @@
 """Teams view."""
+
 import json
 import logging
-from auvsi_suas.models.uas_telemetry import UasTelemetry
 from auvsi_suas.models.takeoff_or_landing_event import TakeoffOrLandingEvent
+from auvsi_suas.models.uas_telemetry import UasTelemetry
 from auvsi_suas.views.decorators import require_superuser
+from auvsi_suas.views.missions import active_mission
 from django.contrib.auth.models import User
 from django.http import HttpResponse
 from django.http import HttpResponseBadRequest
@@ -92,3 +94,34 @@ class TeamsId(View):
 
         return HttpResponse(
             json.dumps(user_json(user)), content_type="application/json")
+
+
+def mission_for_request(request_params):
+    """Gets the mission for the request.
+
+    Args:
+        request_params: The request parameter dict. If this has a 'mission'
+            parameter, it will get the corresponding mission.
+    Returns:
+        Returns (MissionConfig, HttpResponse). The MissionConfig is
+        the one corresponding to the request parameter, or the single active
+        MissionConfig if one exists. The HttpResponse is the appropriate error
+        if a MissionConfig could not be obtained.
+    """
+    # If specific mission requested, get it.
+    if 'mission' in request_params:
+        try:
+            mission_id_str = request_params['mission']
+            mission_id = int(mission_id_str)
+            mission = MissionConfig.objects.get(pk=mission_id)
+            return (mission, None)
+        except ValueError:
+            logger.warning('Invalid mission ID given. ID: %d.', mission_id_str)
+            return (None,
+                    HttpResponseBadRequest('Mission ID is not an integer.'))
+        except MissionConfig.DoesNotExist:
+            logger.warning('Given mission ID not found. ID: %d.', mission_id)
+            return (None, HttpResponseBadRequest('Mission not found.'))
+
+    # Mission not specified, get the single active mission.
+    return active_mission()
